+++ nodeconductor/structure/tests/unittests/test_signals.py
@@ -1,4 +1,6 @@
 from django.test import TestCase
+from django.utils import timezone
+from freezegun import freeze_time
 
 from nodeconductor.structure.tests import factories
 from nodeconductor.structure.tests import models as test_models
@@ -57,3 +59,28 @@ class ServiceProjectLinkSignalsTest(TestCase):
     def link_exists(self, project, service):
         return test_models.TestServiceProjectLink.objects.filter(
             project=project, service=service).exists()
+
+
+class StartTimeTest(TestCase):
+    def test_if_resource_becomes_online_start_time_is_initialized(self):
+        now = timezone.now()
+        with freeze_time(now):
+            vm = factories.TestNewInstanceFactory(runtime_state='in-progress', start_time=None)
+            vm.runtime_state = 'online'
+            vm.save()
+            vm.refresh_from_db()
+            self.assertEqual(vm.start_time, now)
+
+    def test_if_resource_becomes_offline_start_time_is_resetted(self):
+        vm = factories.TestNewInstanceFactory(runtime_state='paused', start_time=timezone.now())
+        vm.runtime_state = 'offline'
+        vm.save()
+        vm.refresh_from_db()
+        self.assertEqual(vm.start_time, None)
+
+    def test_if_runtime_state_changed_to_other_state_start_time_is_not_modified(self):
+        vm = factories.TestNewInstanceFactory(runtime_state='online', start_time=None)
+        vm.runtime_state = 'extending'
+        vm.save()
+        vm.refresh_from_db()
+        self.assertEqual(vm.start_time, None)